Experiment with lightning enhancement of existing websites (#190)
* Experiment with lightning enhancement of existing websites

This inspects websites for lightning information (e.g. a special file in a GitHub repository or the podcast value data on podcast index). This information then can be used to display in the popup and allow the user to easily send a lightning payment or stream payments.

* Add YouTube and Twitter lightning batteries

* Load origin data for monetization tag

* Refactor structure of lightning batteries

* Highlight extension icon when enhancement is available

* Fix LBE_LIGHTNING_DATA return value

we have an array of recipients - currently only one recipient is supported

* feat: show card when lndata available

* feat: button can be an anchor

* feat: call lnurl action

* fix: onClick

* New format for the GitHub .lndonate file

* Correct callable HTTP URL

lnurl:// must be replaced with https to make it callable

* feat: only show donate when there isn't an allowance

* Fix twitter lnurl exraction

there might be some control characters somewhere... so we compare the char codes.

* Cleanup

* Show publisher image

* ups, broke that during merge

* Pass on enhancement origin data to lnurl popup

* Fix metadata

* Wording

I think "Donate" is too specific about "donating" we just want to send sats

* The description hash must be compared as defined by the LNURL spec

* Refactor how we extract the LN data

Not all pages to a full reload (e.g. Twitter) thus we now trigger the extraction from the background script.

* YouTube

* Try to more reliably extract the lnurl/lightning address from a twitter profile

* Fix monetization meta tag extraction

* Load content script at document_end

this means the DOM has finished loading, but resources such as scripts and images may still be loading.

* Fix account switcher

mind the THIS :D

* Better twitter name extraction from the sidebar

We only look for the first span (otherwise "follows-you" is also part of the text)

* Text

* Simplify user description selector

not all users have a header photo

* feat: show lnurlpay loading state

* Fix button import path

* Use new more lightning specific metadata tag

we now use meta name="lightning" content="lnurlp:...."
this is more lightning specific and can still be easily extended.

* Trigger payment metadata extraction only from background script

* Revert "Trigger payment metadata extraction only from background script"

This reverts commit 3295cd5a22538c591d35ecf64a57e7f3d444fa4a.

I think it's broken

* Trigger lightning data extraction from the background script again

There is no real event to know when a page is fully loaded.
we use the tabs.onUpdated "complete" event and add a bit of timeout to hope all async data is fully loaded

* Cleanup

* Ignore the content of the monetization meta tag

* Check for lnurlp in the meta tag again

we only support lnurp for now

* refactor: use interval to grab twitter data

* refactor: less retries and improved data retrieval

* refactor: only grab twitter userdata on specific pages

* refactor: observe dom instead of polling

* LNURLs must start with https

* More stable way to extract the lightning fallback parameter

* Cleanup

* Only use twitter and metatag lightning battery for now

Enable others later after more testing

* cleanup

Co-authored-by: Dylan Companjen <dylancompanjen@gmail.com>